authorization implemented

+++ src/api/songs/handler.js
@@ -18,9 +18,10 @@ class SongsHandler {
       const {
         title = "untitled", year, performer, genre, duration,
       } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
       const songId = await this._service.addSong({
-        title, year, performer, genre, duration,
+        title, year, performer, genre, duration, owner: credentialId,
       });
 
       const response = h.response({
@@ -53,9 +54,10 @@ class SongsHandler {
     }
   }
 
-  async getSongsHandler(_request, h) {
+  async getSongsHandler(request, h) {
     try {
-      const songs = await this._service.getSongs();
+      const { id: credentialId } = request.auth.credentials;
+      const songs = await this._service.getSongs(credentialId);
       const songsProps = songs.map((song) => ({
         id: song.id,
         title: song.title,
@@ -91,7 +93,11 @@ class SongsHandler {
   async getSongByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
       const song = await this._service.getSongById(id);
+
       return {
         status: "success",
         data: {
@@ -123,7 +129,9 @@ class SongsHandler {
     try {
       this._validator.validateSongPayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
+      await this._service.verifySongOwner(id, credentialId);
       await this._service.editSongById(id, request.payload);
 
       return {
@@ -154,6 +162,9 @@ class SongsHandler {
   async deleteSongByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
       await this._service.deleteSongById(id);
       return {
         status: "success",
